* code runs now, but give NaN at some position (that's new...)

+++ include/testdata_1.h
@@ -36,10 +36,10 @@
 						3,	0,	1,
 						2,	4,	1,
 						4,	3,	1 };
-	int host_z_mesh = 1;
+	double host_z_mesh = 1;
 	int host_size_t = 4;
 	int host_size_p = 5;
-	double host_mesh_z = 1;
+	int host_mesh_z = 1;
 	
 	double host_clad_abs = 5.5;
 	int host_clad_num = 3;
